BSYS add .py script

+++ S3/BS/BSCode/c19_tlb/plot_access.py
@@ -15,8 +15,8 @@ maxNumPages = 16384
 numPages = 1
 trials = 1000000
 
-arrayNumPages = np.zeros(round(np.log(maxNumPages) / np.log(2)))
-arrayTime = np.zeros(round(np.log(maxNumPages) / np.log(2)))
+arrayNumPages = np.zeros(int(round(np.log(maxNumPages) / np.log(2))))
+arrayTime = np.zeros(int(round(np.log(maxNumPages) / np.log(2))))
 
 count = 0
